RLB: Wait for uploader on shutdown using new on_shutdown mechanism
This is a first prototype.
It adds a new callback `on_shutdown`,
which gets called just after we drain and shut down the dispatcher,
but before Glean is considered fully shutdown.

The callback MUST NOT put any new tasks on the dispatcher.
The callback SHOULD NOT block arbitrarily long.

The RLB implementation of that callback instructs a potential uploader
thread to end operation after it finished one upload.
It also blocks any new upload thread from getting started.
This means an in-flight upload can finish as usual, but no new one
is launched.

This also adds a new test.
The only way to reliably test this is from the outside,
so the test invokes an example implementation and checks that the
pending_pings directory is empty after the program ends.
The uploader inside sleeps for a short period of time.
Without waiting for this uploader the main thread would end and kill the
uploader thread before it wakes up.
With waiting the uploader finishes correctly before the main thread
exits.

The downsides as I see them:

* It's another callback, with very strict requirements. We can't check
  any of them.
* This is always called on shutdown and introduces a potential arbitrary
  blocking.
  We can't enforce proper timeouts on this and there's no way to signal
  the uploader thread to not do too much anymore..
* I dislike the 3-state integer of `thread_running`,
  but afaik it's the best we can do right now.
* It's only done for RLB and it's not clear how others would even use
  this, they generally don't use shutdown anyway.